pyproject.toml
[project]
name = "yourcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coloredlogs>=15.0.1",
    "markdownify>=1.1.0",
    "playwright>=1.51.0",
    "playwright-stealth>=1.0.6",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-slugify>=8.0.4",
    "requests>=2.32.3",
    "setuptools>=79.0.1",
]

[tool.ruff]
line-length = 160
